pyproject.toml
[project]
name = "genai-learning"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.37.1",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
]
